# docs/openapi.json
{
  "components": {
    "schemas": {
      "ApprovalDecision": {
        "properties": {
          "approved_by": {
            "default": "cantor",
            "title": "Approved By",
            "type": "string"
          },
          "note": {
            "default": "",
            "title": "Note",
            "type": "string"
          }
        },
        "title": "ApprovalDecision",
        "type": "object"
      },
...
      "DiscoverRequest": {
        "properties": {
          "goal": {
            "minLength": 1,
            "title": "Goal",
            "type": "string"
          },
          "limit": {
            "default": 10,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          }
        },
        "required": [
          "goal"
        ],
        "title": "DiscoverRequest",
        "type": "object"
      },
      "DiscoverResponse": {
        "properties": {
          "contract_version": {
            "default": "1.0",
            "title": "Contract Version",
            "type": "string"
          },
          "goal": {
            "title": "Goal",
            "type": "string"
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/CapabilityMatch"
            },
            "title": "Matches",
            "type": "array"
          }
        },
        "required": [
          "goal"
        ],
        "title": "DiscoverResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobRequest": {
        "properties": {
          "approval_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Id"
          },
          "goal": {
            "default": "",
            "title": "Goal",
            "type": "string"
          },
          "inputs": {
            "additionalProperties": true,
            "title": "Inputs",
            "type": "object"
          },
          "policy": {
            "$ref": "#/components/schemas/Policy"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "requested_by": {
            "default": "echo",
            "title": "Requested By",
            "type": "string"
          },
          "skill": {
            "title": "Skill",
            "type": "string"
          }
        },
        "required": [
          "skill",
          "provider"
        ],
        "title": "JobRequest",
        "type": "object"
      },
      "PlanCreateRequest": {
        "properties": {
          "goal": {
            "minLength": 1,
            "title": "Goal",
            "type": "string"
          },
          "inputs": {
            "additionalProperties": true,
            "title": "Inputs",
            "type": "object"
          }
        },
        "required": [
          "goal"
        ],
        "title": "PlanCreateRequest",
        "type": "object"
      },
...
      "PlanEventsResponse": {
        "properties": {
          "contract_version": {
            "default": "1.0",
            "title": "Contract Version",
            "type": "string"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/PlanEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "plan_id": {
            "title": "Plan Id",
            "type": "string"
          }
        },
        "required": [
          "plan_id"
        ],
        "title": "PlanEventsResponse",
        "type": "object"
      },
      "PlanExecutionAccepted": {
        "properties": {
          "contract_version": {
            "default": "1.0",
            "title": "Contract Version",
            "type": "string"
          },
          "plan_id": {
            "title": "Plan Id",
            "type": "string"
          },
          "status": {
            "const": "running",
            "default": "running",
            "title": "Status",
            "type": "string"
          },
          "step_jobs": {
            "items": {
              "$ref": "#/components/schemas/PlanStepJob"
            },
            "title": "Step Jobs",
            "type": "array"
          }
        },
        "required": [
          "plan_id"
        ],
        "title": "PlanExecutionAccepted",
        "type": "object"
      },
      "PlanExplanation": {
        "properties": {
          "contract_version": {
            "default": "1.0",
            "title": "Contract Version",
            "type": "string"
          },
          "goal": {
            "title": "Goal",
            "type": "string"
          },
          "plan_id": {
            "title": "Plan Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/PlanStepExplanation"
            },
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "plan_id",
          "goal",
          "status"
        ],
        "title": "PlanExplanation",
        "type": "object"
      },
...
      "RejectionDecision": {
        "properties": {
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "rejected_by": {
            "default": "cantor",
            "title": "Rejected By",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "RejectionDecision",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Canto",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/approvals/{approval_id}/approve": {
      "post": {
        "operationId": "approve_approvals__approval_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Approve Approvals  Approval Id  Approve Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve"
      }
    },
    "/approvals/{approval_id}/reject": {
      "post": {
        "operationId": "reject_approvals__approval_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectionDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Reject Approvals  Approval Id  Reject Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject"
      }
    },
    "/discover": {
      "post": {
        "operationId": "discover_discover_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoverRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoverResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discover"
      }
    },
...
    "/jobs": {
      "post": {
        "operationId": "create_job_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Job Jobs Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Job"
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Job Jobs  Job Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job"
      }
    },
    "/jobs/{job_id}/artifacts": {
      "get": {
        "operationId": "get_artifacts_jobs__job_id__artifacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Artifacts Jobs  Job Id  Artifacts Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Artifacts"
      }
    },
    "/jobs/{job_id}/artifacts/{artifact_name}": {
      "get": {
        "operationId": "get_artifact_jobs__job_id__artifacts__artifact_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_name",
            "required": true,
            "schema": {
              "title": "Artifact Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Artifact Jobs  Job Id  Artifacts  Artifact Name  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Artifact"
      }
    },
    "/jobs/{job_id}/events": {
      "get": {
        "operationId": "get_events_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Events Jobs  Job Id  Events Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Events"
      }
    },
    "/jobs/{job_id}/promote": {
      "post": {
        "operationId": "promote_job_jobs__job_id__promote_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Promote Job Jobs  Job Id  Promote Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Promote Job"
      }
    },
    "/jobs/{job_id}/recover": {
      "post": {
        "operationId": "recover_job_jobs__job_id__recover_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Recover Job Jobs  Job Id  Recover Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recover Job"
      }
    },
    "/plans": {
      "post": {
        "operationId": "create_plan_plans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionPlan"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Plan"
      }
    },
    "/plans/{plan_id}": {
      "get": {
        "operationId": "get_plan_plans__plan_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionPlan"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Plan"
      }
    },
    "/plans/{plan_id}/approve": {
      "post": {
        "operationId": "approve_plan_plans__plan_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionPlan"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Plan"
      }
    },
    "/plans/{plan_id}/events": {
      "get": {
        "operationId": "plan_events_plans__plan_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanEventsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Plan Events"
      }
    },
    "/plans/{plan_id}/execute": {
      "post": {
        "operationId": "execute_plan_plans__plan_id__execute_post",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanExecutionAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Plan"
      }
    },
    "/plans/{plan_id}/explain": {
      "get": {
        "operationId": "explain_plan_plans__plan_id__explain_get",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanExplanation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Explain Plan"
      }
    },
...
    "/skills/{skill_name}": {
      "get": {
        "operationId": "get_skill_skills__skill_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "skill_name",
            "required": true,
            "schema": {
              "title": "Skill Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Skill Skills  Skill Name  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Skill"
      }
    },
    "/skills/{skill_name}/providers/{provider_name}": {
      "get": {
        "operationId": "get_provider_skills__skill_name__providers__provider_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "skill_name",
            "required": true,
            "schema": {
              "title": "Skill Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "provider_name",
            "required": true,
            "schema": {
              "title": "Provider Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Provider Skills  Skill Name  Providers  Provider Name  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Provider"
      }
    }
  }
}